Initial Contribution to CBC Order Processing System
Uploaded the backend and frontend components of the CBC Order Processing System, reflecting my contributions to the project. This includes the implementation of key Java classes, controllers, and repositories on the backend, as well as the HTML/CSS interface for the frontend. The project is a part of my coursework at London Metropolitan University and demonstrates my involvement in designing, implementing, and testing individual components of the system.